move tabu_list to a new file
